Crate rust_stix2

#rust-stix2

An implementation of STIX 2.1 in Rust.

This is a work in progress implementation of the object and features of STIX 2.1 as a Rust library. Eventually, this will become a full Rust-based STIX API.

The objects and features in this library are intended to conform to the STIX 2.1 standards detailed in this document.

Modules

base

A collection of STIX wide properties and features that could be used by any STIX object or type.

bundles

Contains the implementation logic for STIX Bundles

custom_objects

Contains the implementation logic for unrecognized custom STIX Objects.

cyber_observable_objects

Data structures and functions for implementing STIX Cyber-observable Objects (SCOs).

datastore

The datastore is a top-level component that includes filesystem, filters, structs, and enums.

domain_objects

Data structures and functions for implementing STIX Domain Objects (SDOs).

error

A custom STIX Error type, with its associated functions.

extensions

Contains the data structures and implementation for predefined Stix Cyber-observable object extensions.

json

Functions for deserializing STIX objects from JSON strings.

meta_objects

Data structures and functions for implementing STIX Meta Objects (SMOs).

object

Top level STIX object structures and implementation

pattern

A Rust implementaiton of STIX Patterning grammar that use the nom combinator parser

relationship_objects

Data structures and functions for implementing STIX Relationship Objects (SROs).

types

Core types for representing and manipulating STIX data structures.