rust_stix2::base
Struct CommonPropertiesBuilder

pub struct CommonPropertiesBuilder {
    pub builder_type: BuilderType,
    pub properties: CommonProperties,
    /* private fields */
}

Builder struct for common STIX properties.

This follows the โRust builder pattern,โ where we use a new() function to construct a Builder with a minimum set of required fields, then set additional fields with their own setter functions. Once all fields have been set, the build() function will take all of the fields in the Builder struct and use them to create the final CommonProperties struct.

This struct is intended to be nested inside of a STIX Objectโs Builder struct to help with building that object by constructing the common properties.

Because different STIX Objects have different common properties, they should check that they have all required properties and do not have any omitted properties as part of their own stix_check() function during the build stage.

Fields
builder_type: BuilderType

Whether we are creating a new object or versioning an existing one.

properties: CommonProperties

The common properties that will be added to the object being built

Implementations

impl CommonPropertiesBuilder

pub fn new( object_name: &str, type_name: &str, ) -> Result<CommonPropertiesBuilder, Error>

Construct a CommonPropertiesBuilder with default values for the properties Used when building a new STIX Object.

pub fn version( object_name: &str, old: &CommonProperties, ) -> Result<CommonPropertiesBuilder, Error>

Construct a CommonPropertiesBuilder by cloning an existing set of properites Used when versioning an existing STIX Object.

pub fn created_by_ref(self, id: Identifier) -> Result<Self, Error>

Set the created_by_ref field for an object under construction This is only allowed when creating a new object, not when versioning an existing one, as only the original creator of an object can version it.

pub fn labels(self, labels: Vec<String>) -> Self

Set the optional labels field for an object under construction.

pub fn confidence(self, confidence: u8) -> Self

Set the optional confidence field for an object under construction.

pub fn lang(self, language: String) -> Self

Set the optional lang field for an object under construction. If the language is English (โenโ), this does not need to be set (but it can be if specificity is desired).

pub fn external_references(self, references: Vec<ExternalReference>) -> Self

Set the optional external_references field for an object under construction.

pub fn object_marking_refs(self, references: Vec<Identifier>) -> Self

Set the optional object_marking_refs field for an object under construction.

pub fn defanged(self) -> Self

Set the optional defanged field to Some(true) for an object under construction.

pub fn granular_markings(self, markings: Vec<GranularMarking>) -> Self

Set the optional granular_markings field for an object under construction.

pub fn add_extension( self, key: &str, extension: StixDictionary<DictionaryValue>, ) -> Result<Self, Error>

Add an optional extension to the extensions field for an object under construction, creating the field if it does not already exist.

pub fn build(&self) -> CommonProperties