rust_stix2::cyber_observable_objects::sco
Struct CyberObject

pub struct CyberObject {
    pub object_type: CyberObjectType,
    pub common_properties: CommonProperties,
}

A STIX Cyber-observable Object (SCO) of some type.

STIX defines a set of STIX Cyber-observable Objects (SCOs) for characterizing host-based and network-based information.

SCOs are used by various STIX Domain Objects (SDOs) to provide supporting context. The Observed Data SDO, for example, indicates that the raw data was observed at a particular time.

For more information see https://docs.oasis-open.org/cti/stix/v2.1/os/stix-v2.1-os.html#_rosvg2qjx4h4

Fields
object_type: CyberObjectType

Identifies the SCO type of SCO.

common_properties: CommonProperties

Common object properties

Implementations

impl CyberObject

pub fn from_json(json: &str, allow_custom: bool) -> Result<Self, Error>

Deserializes an SCO from a JSON String. Checks that all fields conform to the STIX 2.1 standard If the allow_custom flag is flase, checks that there are no fields in the JSON String that are not in the SRO type definition

pub fn is_revoked(&self) -> bool