rust_stix2::cyber_observable_objects::sco
Struct CyberObjectBuilder

pub struct CyberObjectBuilder { /* private fields */ }

Builder struct for SCOs.

This follows the “Rust builder pattern,” where we use a new() function to construct a Builder with a minimum set of required fields, then set additional fields with their own setter functions. Once all fields have been set, the build() function will take all of the fields in the Builder struct and use them to create the final CyberObject struct.

Note: Because different types of SDOs have different fields, and fields present in multiple types may be requried in some and optional in others, it is possible for a CyberObjectBuilder to be in an incomplete state prior to constructing a CyberObject. The build() function will error in such a case.

Implementations

impl CyberObjectBuilder

pub fn belongs_to_ref(self, belongs_to_ref: Identifier) -> Result<Self, Error>

Set the belongs_to_ref field for an SCO, if the SCO is of a type that has that field

pub fn new(type_name: &str) -> Result<CyberObjectBuilder, Error>

Creates a new STIX 2.1 CyberObjectBuilder of the given type

The type field must be lowercase, with words separated by -

Automatically generates an id Other fields are set to their Default (which is `None`` for optional fields)

pub fn created_by_ref(self, id: Identifier) -> Result<Self, Error>

Set the optional created_by_ref field for an SCO under construction. This is only allowed when creating a new SCO, not when versioning an existing one, as only the original creator of an object can version it.

pub fn labels(self, labels: Vec<String>) -> Self

Set the optional labels field for an SCO under construction.

pub fn confidence(self, confidence: u8) -> Self

Set the optional confidence field for an SCO under construction.

pub fn lang(self, language: String) -> Self

Set the optional lang field for an SCO under construction. If the language is English (“en”), this does not need to be set (but it can be if specificity is desired).

pub fn external_references(self, references: Vec<ExternalReference>) -> Self

Set the optional external_references field for an SCO under construction.

pub fn object_marking_refs(self, references: Vec<Identifier>) -> Self

Set the optional object_marking_refs field for an SCO under construction.

pub fn granular_markings(self, markings: Vec<GranularMarking>) -> Self

Set the optional granular_markings field for an SCO under construction.

pub fn add_extension( self, key: &str, extension: StixDictionary<DictionaryValue>, ) -> Result<Self, Error>

Add an optional extension to the extensions field for an SCO under construction, creating the field if it does not exist

pub fn additional_header_fields( self, additional_header_fields: StixDictionary<Vec<String>>, ) -> Result<Self, Error>

Set the optional additional_header_fields property for an SCO under construction

pub fn account_created(self, account_created: String) -> Result<Self, Error>

Set the account_created field for an SCO under construction.

pub fn account_expires(self, account_expires: String) -> Result<Self, Error>

Set the account_expires field for an SCO under construction.

pub fn account_first_login( self, account_first_login: String, ) -> Result<Self, Error>

Set the account_first_login field for an SCO under construction.

pub fn account_last_login( self, account_last_login: String, ) -> Result<Self, Error>

Set the account_last_login field for an SCO under construction.

pub fn account_login(self, account_login: String) -> Result<Self, Error>

Set the account_login field for an Sso under construction.

pub fn account_type(self, account_type: String) -> Result<Self, Error>

Set the account_type field for an Sso under construction.

pub fn atime(self, atime: &str) -> Result<Self, Error>

Set the optional atime property for an SCO under construction

pub fn bcc_refs(self, bcc_refs: Vec<Identifier>) -> Result<Self, Error>

Set the optional bcc_refs property for an SCO under construction

pub fn belongs_to_refs(self, refs: Vec<Identifier>) -> Result<Self, Error>

Set the optional belongs_to_refs property for an SCO under construction

pub fn body(self, body: String) -> Result<Self, Error>

Set the optional body property for an SCO under construction

pub fn body_multipart( self, body_multipart: Vec<EmailMimeCompomentType>, ) -> Result<Self, Error>

Set the optional body_multipart property for an SCO under construction

pub fn cc_refs(self, cc_refs: Vec<Identifier>) -> Result<Self, Error>

Set the optional cc_refs property for an SCO under construction

pub fn can_escalate_privs(self, can_escalate_privs: bool) -> Result<Self, Error>

Set the can_escalate_privs field for an SCO under construction.

pub fn child_refs(self, child_refs: Vec<Identifier>) -> Result<Self, Error>

Set the optional child_refs property for an SCO under construction

pub fn command_line(self, command_line: String) -> Result<Self, Error>

Set the optional command_line property for an SCO under construction

pub fn contains_refs( self, contains_refs: Vec<Identifier>, ) -> Result<Self, Error>

Set the optional contains_refs property for an SCO under construction

pub fn content_ref(self, content_ref: Identifier) -> Result<Self, Error>

Set the optional content_ref property for an SCO under construction

pub fn content_type(self, content_type: String) -> Result<Self, Error>

Set the optional content_type property for an SCO under construction

pub fn credential_last_changed( self, credential_last_changed: String, ) -> Result<Self, Error>

Set the optional credential_last_changed field for an SCO under construction.

pub fn cpe(self, cpe: String) -> Result<Self, Error>

Set the optional cpe field for an sco under construction.

pub fn created_time(self, created_time: Timestamp) -> Result<Self, Error>

Set the optional created_time property for an SCO under construction

pub fn creator_user_ref( self, creator_user_ref: Identifier, ) -> Result<Self, Error>

Set the optional creator_user_ref property for an SCO under construction

pub fn credential(self, credential: String) -> Result<Self, Error>

Set the optional credential field for an Sso under construction.

pub fn ctime(self, ctime: &str) -> Result<Self, Error>

Set the optional ctime property for an SCO under construction

pub fn cwd(self, cwd: String) -> Result<Self, Error>

Set the optional cwd property for an SCO under construction

pub fn data(self, data: String) -> Result<Self, Error>

Set the optional data property for an SCO under construction

pub fn date(self, date: Timestamp) -> Result<Self, Error>

Set the optional date property for an SCO under construction

pub fn data_type(self, data_type: String) -> Result<Self, Error>

Set the optional data_type property for an SCO under construction

pub fn decryption_key(self, decryption_key: String) -> Result<Self, Error>

Set the optional description_key property for an SCO under construction

pub fn display_name(self, display_name: String) -> Result<Self, Error>

Set the optional display_name property for an SCO under construction

pub fn dst_byte_count(self, dst_byte_count: u64) -> Result<Self, Error>

Set the optional dst_byte_count property for an SCO under construction

pub fn dst_packets(self, dst_packets: u64) -> Result<Self, Error>

Set the optional dst_packets property for an SCO under construction

pub fn dst_payload_ref(self, dst_payload_ref: Identifier) -> Result<Self, Error>

pub fn dst_port(self, dst_port: u64) -> Result<Self, Error>

pub fn dst_ref(self, dst_ref: Identifier) -> Result<Self, Error>

pub fn encapsulated_by_ref( self, encapsulated_by_ref: Identifier, ) -> Result<Self, Error>

pub fn encapsulates_refs( self, encapsulates_refs: Vec<Identifier>, ) -> Result<Self, Error>

pub fn encryption_algorithm( self, encryption_algorithm: EncryptionAlgorithm, ) -> Result<Self, Error>

pub fn end(self, end: Timestamp) -> Result<Self, Error>

pub fn environment_variables( self, environment_variables: StixDictionary<Vec<String>>, ) -> Result<Self, Error>

pub fn from_ref(self, from_ref: Identifier) -> Result<Self, Error>

pub fn hashes(self, hashes: Hashes) -> Result<Self, Error>

pub fn ipfix( self, ipfix: StixDictionary<DictionaryValue>, ) -> Result<Self, Error>

pub fn is_active(self, is_active: bool) -> Result<Self, Error>

pub fn is_disabled(self, is_disabled: bool) -> Result<Self, Error>

Set the is_disabled field for an Sso under construction.

pub fn is_hidden(self, is_hidden: bool) -> Result<Self, Error>

pub fn is_self_signed(self, is_self_signed: bool) -> Result<Self, Error>

pub fn issuer(self, issuer: String) -> Result<Self, Error>

pub fn image_ref(self, image_ref: Identifier) -> Result<Self, Error>

pub fn is_multipart(self) -> Result<Self, Error>

pub fn is_privileged(self, is_privileged: bool) -> Result<Self, Error>

Set the is_privileged field for an Sso under construction.

pub fn is_service_account(self, is_service_account: bool) -> Result<Self, Error>

Set the is_service_account field for an Sso under construction.

pub fn key(self, key: String) -> Result<Self, Error>

pub fn languages(self, languages: Vec<String>) -> Result<Self, Error>

Set the languages field for an Sso under construction.

pub fn magic_number_hex(self, magic_number_hex: String) -> Result<Self, Error>

pub fn message_id(self, message_id: String) -> Result<Self, Error>

pub fn mime_type(self, mime_type: String) -> Result<Self, Error>

pub fn modified_time(self, modified_time: Timestamp) -> Result<Self, Error>

pub fn mtime(self, mtime: &str) -> Result<Self, Error>

pub fn name(self, name: String) -> Result<Self, Error>

Set the name field for an sco under construction.

pub fn name_enc(self, name_enc: String) -> Result<Self, Error>

pub fn number(self, number: u64) -> Result<Self, Error>

pub fn number_of_subkeys(self, number_of_subkeys: i64) -> Result<Self, Error>

pub fn opened_connection_refs( self, opened_connection_refs: Vec<Identifier>, ) -> Result<Self, Error>

pub fn parent_directory_ref( self, parent_directory_ref: Identifier, ) -> Result<Self, Error>

pub fn parent_ref(self, parent_ref: Identifier) -> Result<Self, Error>

pub fn path(self, path: String) -> Result<Self, Error>

pub fn path_enc(self, path_enc: String) -> Result<Self, Error>

pub fn payload_bin(self, payload_bin: String) -> Result<Self, Error>

pub fn pid(self, pid: i64) -> Result<Self, Error>

pub fn protocols(self, protocols: Vec<String>) -> Result<Self, Error>

pub fn raw_email_ref(self, raw_email_ref: Identifier) -> Result<Self, Error>

pub fn recieved_lines(self, recieved_lines: Vec<String>) -> Result<Self, Error>

pub fn resolves_to_refs(self, refs: Vec<Identifier>) -> Result<Self, Error>

pub fn rir(self, rir: String) -> Result<Self, Error>

pub fn sender_ref(self, sender_ref: Identifier) -> Result<Self, Error>

pub fn serial_number(self, serial_number: String) -> Result<Self, Error>

pub fn signature_algorithm( self, signature_algorithm: String, ) -> Result<Self, Error>

pub fn size(self, size: u64) -> Result<Self, Error>

pub fn src_byte_count(self, src_byte_count: u64) -> Result<Self, Error>

pub fn src_packets(self, src_packets: u64) -> Result<Self, Error>

pub fn src_payload_ref(self, src_payload_ref: Identifier) -> Result<Self, Error>

pub fn src_ref(self, src_ref: Identifier) -> Result<Self, Error>

pub fn src_port(self, src_port: u64) -> Result<Self, Error>

pub fn start(self, start: Timestamp) -> Result<Self, Error>

pub fn subject(self, subject: String) -> Result<Self, Error>

pub fn subject_public_key_algorithm( self, subject_public_key_algorithm: String, ) -> Result<Self, Error>

pub fn subject_public_key_exponent( self, subject_public_key_exponent: i64, ) -> Result<Self, Error>

pub fn subject_public_key_modulus( self, subject_public_key_modulus: String, ) -> Result<Self, Error>

pub fn swid(self, swid: String) -> Result<Self, Error>

Set the swid field for an sco under construction.

pub fn to_refs(self, to_refs: Vec<Identifier>) -> Result<Self, Error>

pub fn url(self, url_str: String) -> Result<Self, Error>

pub fn value(self, value: String) -> Result<Self, Error>

pub fn values(self, values: Vec<WindowsRegistryKeyType>) -> Result<Self, Error>

pub fn vendor(self, vendor: String) -> Result<Self, Error>

Set the vendor field for an sco under construction.

pub fn user_id(self, user_id: String) -> Result<Self, Error>

Set the user_id field for an Sco under construction.

pub fn validity_not_after( self, validity_not_after: Timestamp, ) -> Result<Self, Error>

pub fn validity_not_before( self, validity_not_before: Timestamp, ) -> Result<Self, Error>

pub fn version(self, version: String) -> Result<Self, Error>

Set the version field for an sco under construction.

pub fn x509_v3_extensions( self, x509_v3_extensions: X509V3Extensions, ) -> Result<Self, Error>

pub fn build(self) -> Result<CyberObject, Error>

Builds a new SCO, using the information found in the DomainObjectBuilder

This performs a final check that all required fields for a given SDO type are included before construction. If possible, it generates a UUIDv5 for the SCO, in place of a UUIDv4. This also runs the stick_check() validation method on the newly constructed SDO.