rust_stix2::cyber_observable_objects::sco_types
Struct X509V3Extensions

pub struct X509V3Extensions {

    pub basic_constraints: Option<String>,
    pub name_constraints: Option<String>,
    pub policy_constraints: Option<String>,
    pub key_usage: Option<String>,
    pub extended_key_usage: Option<String>,
    pub subject_key_identifier: Option<String>,
    pub authority_key_identifier: Option<String>,
    pub subject_alternative_name: Option<String>,
    pub issuer_alternative_name: Option<String>,
    pub subject_directory_attributes: Option<String>,
    pub crl_distribution_points: Option<String>,
    pub inhibit_any_policy: Option<String>,
    pub private_key_usage_period_not_before: Option<Timestamp>,
    pub private_key_usage_period_not_after: Option<Timestamp>,
    pub certificate_policies: Option<String>,
    pub policy_mappings: Option<String>,

}

X.509 v3 Extensions

The X.509 v3 Extensions object represents the properties of the X.509 v3 extensions that may be used in the certificate.

For more information see https://docs.oasis-open.org/cti/stix/v2.1/os/stix-v2.1-os.html#_oudvonxzdlku

Fields
basic_constraints: Option<String>

Specifies a multi-valued extension which indicates whether a certificate is a CA certificate.

name_constraints: Option<String>

Specifies a namespace within which all subject names in subsequent certificates in a certification path MUST be located.

policy_constraints: Option<String>

Specifies any constraints on path validation for certificates issued to CAs.

key_usage: Option<String>

Specifies a multi-valued extension consisting of a list of names of the permitted key usages.

extended_key_usage: Option<String>

Specifies a list of usages indicating purposes for which the certificate public key can be used for.

subject_key_identifier: Option<String>

Specifies the identifier that provides a means of identifying certificates that contain a particular public key.

authority_key_identifier: Option<String>

Specifies the identifier that provides a means of identifying the public key corresponding to the private key used to sign a certificate.

subject_alternative_name: Option<String>

Specifies the additional identities to be bound to the subject of the certificate.

issuer_alternative_name: Option<String>

Specifies the additional identities to be bound to the issuer of the certificate.

subject_directory_attributes: Option<String>

Specifies the identification attributes (e.g., nationality) of the subject.

crl_distribution_points: Option<String>

Specifies how CRL information is obtained.

inhibit_any_policy: Option<String>

Specifies the number of additional certificates that may appear in the path before anyPolicy is no longer permitted.

private_key_usage_period_not_before: Option<Timestamp>

Specifies the date on which the validity period begins for the private key, if it is different from the validity period of the certificate.

private_key_usage_period_not_after: Option<Timestamp>

Specifies the date on which the validity period ends for the private key, if it is different from the validity period of the certificate.

certificate_policies: Option<String>

Specifies a sequence of one or more policy information terms, each of which consists of an object identifier (OID) and optional qualifiers.

policy_mappings: Option<String>

Specifies one or more pairs of OIDs; each pair includes an issuerDomainPolicy and a subjectDomainPolicy.