rust_stix2::cyber_observable_objects
Module vocab

Enumerations for standardizing and categorizing types for SCOs. Enumerations for standardizing and categorizing types for SCOs.

Structs

AccountTypeVocabularyIter

An iterator over the variants of AccountTypeVocabulary

EncryptionAlgorithmIter

An iterator over the variants of EncryptionAlgorithm

IanaServiceNamesEnumIter

An iterator over the variants of IanaServiceNamesEnum

NetworkSocketAddressFamilyEnumIter

An iterator over the variants of NetworkSocketAddressFamilyEnum

NetworkSocketTypeEnumIter

An iterator over the variants of NetworkSocketTypeEnum

WindowsIntegrityEnumIter

An iterator over the variants of WindowsIntegrityEnum

WindowsRegistryDataTypeEnumIter

An iterator over the variants of WindowsRegistryDataTypeEnum

WindowsServiceStartTypeEnumIter

An iterator over the variants of WindowsServiceStartTypeEnum

WindowsServiceStatusEnumIter

An iterator over the variants of WindowsServiceStatusEnum

WindowsServiceTypeEnumIter

An iterator over the variants of WindowsServiceTypeEnum

Enums

AccountTypeVocabulary

Represents the different types and functions of malware within the STIX framework.

EncryptionAlgorithm

The encryption algorithm enumeration

IanaServiceNamesEnum

Internet Assigned Numbers Authority enumeration

NetworkSocketAddressFamilyEnum

Network Socket Address Family Enumeration

NetworkSocketTypeEnum

Network Socket Type Enumeration

WindowsIntegrityEnum

The Windows integrity level enumeration

WindowsRegistryDataTypeEnum

The Windows registry datatype enumeration The Windows registry datatype enumeration

WindowsServiceStartTypeEnum

The Windows service start type enumeration

WindowsServiceStatusEnum

The Windows service status enumeration

WindowsServiceTypeEnum

The Windows service type enumeration