rust_stix2::domain_objects::sdo
Struct DomainObjectBuilder

pub struct DomainObjectBuilder { /* private fields */ }

Builder struct for SDOs.

This follows the “Rust builder pattern,” where we use a new() function to construct a Builder with a minimum set of required fields, then set additional fields with their own setter functions. Once all fields have been set, the build() function will take all of the fields in the Builder struct and use them to create the final DomainObject struct.

Note: Because different types of SDOs have different fields, and fields present in multiple types may be requried in some and optional in others, it is possible for a DomainObjectBuilder to be in an incomplete state prior to constructing a DomainObject. The build() function will error in such a case.

Implementations

impl DomainObjectBuilder

pub fn new(type_name: &str) -> Result<DomainObjectBuilder, Error>

Creates a new STIX 2.1 DomainObjectBuilder of the given type

The type field must be lowercase, with words separated by -

Automatically generates an id Other fields are set to their Default (which is `None`` for optional fields)

pub fn version(old: &DomainObject) -> Result<DomainObjectBuilder, Error>

Create a new STIX 2.1 DomainObjectBuilder by cloning the fields from an existing DomainObject When built, this will create DomainObject as a newer version of the original object.

pub fn created_by_ref(self, id: Identifier) -> Result<Self, Error>

Set the optional created_by_ref field for an SDO under construction. This is only allowed when creating a new SDO, not when versioning an existing one, as only the original creator of an object can version it.

pub fn labels(self, labels: Vec<String>) -> Self

Set the optional labels field for an SDO under construction.

pub fn confidence(self, confidence: u8) -> Self

Set the optional confidence field for an SDO under construction.

pub fn lang(self, language: String) -> Self

Set the optional lang field for an SDO under construction. If the language is English (“en”), this does not need to be set (but it can be if specificity is desired).

pub fn external_references(self, references: Vec<ExternalReference>) -> Self

Set the optional external_references field for an SDO under construction.

pub fn object_marking_refs(self, references: Vec<Identifier>) -> Self

Set the optional object_marking_refs field for an SDO under construction.

pub fn granular_markings(self, markings: Vec<GranularMarking>) -> Self

Set the optional granular_markings field for an SDO under construction.

pub fn add_extension( self, key: &str, extension: StixDictionary<DictionaryValue>, ) -> Result<Self, Error>

Add an optional extension to the extensions field for an SDO under construction, creating the field if it does not exist

pub fn administrative_area(self, area: String) -> Result<Self, Error>

pub fn aliases(self, aliases: Vec<String>) -> Result<Self, Error>

Set the aliases field for an SDO under construction, if the SDO is of a type that has that field

pub fn analysis_definition_version( self, analysis_definition_version: String, ) -> Result<Self, Error>

pub fn analysis_ended(self, analysis_ended: String) -> Result<Self, Error>

pub fn analysis_engine_version( self, analysis_engine_version: String, ) -> Result<Self, Error>

pub fn analysis_sco_refs( self, analysis_sco_refs: Vec<Identifier>, ) -> Result<Self, Error>

pub fn architecture_execution_envs( self, architecture_execution_envs: Vec<String>, ) -> Result<Self, Error>

Set the architecture_execution_envs field for an SDO under construction, if the SDO is of a type that has that field

pub fn authors(self, authors: Vec<String>) -> Result<Self, Error>

Set the authors field for an SDO under construction, if the SDO is of a type that has that field

pub fn capabilities(self, capabilities: Vec<String>) -> Result<Self, Error>

Set the capabilities field for an SDO under construction, if the SDO is of a type that has that field

pub fn city(self, city: String) -> Result<Self, Error>

pub fn configuration_version( self, configuration_version: String, ) -> Result<Self, Error>

Set the configuration_version field for an SDO under construction, if the SDO is of a type that has that field

pub fn contact_information( self, contact_information: String, ) -> Result<Self, Error>

Set the `contact_information (field for an SDO under construction, if the SDO is of a type that has that field

pub fn content(self, content: String) -> Result<Self, Error>

Set the content field for an SDO under construction, if the SDO is of a type that has that field

pub fn context(self, context: String) -> Result<Self, Error>

Set the context field for an SDO under construction, if the SDO is of a type that has that field

pub fn country(self, country: String) -> Result<Self, Error>

pub fn description(self, description: String) -> Result<Self, Error>

Set the description field for an SDO under construction, if the SDO is of a type that has that field

pub fn explanation(self, explanation: String) -> Result<Self, Error>

Set the explanation field for an SDO under construction, if the SDO is of a type that has that field

pub fn first_observed(self, first_observed: String) -> Result<Self, Error>

Set the first_observed field for an SDO under construction, if the SDO is of a type that has that field

pub fn first_seen(self, first_seen: String) -> Result<Self, Error>

Set the first_seen field for an SDO under construction, if the SDO is of a type that has that field

pub fn goals(self, goals: Vec<String>) -> Result<Self, Error>

Set the roles field for an SDO under construction, if the SDO is of a type that has that field

pub fn host_vm_ref(self, host_vm_ref: Identifier) -> Result<Self, Error>

Set the host_vm_ref field for an SDO under construction, if the SDO is of a type that has that field

pub fn identity_class(self, identity_class: String) -> Result<Self, Error>

Set the identity_class field for an SDO under construction, if the SDO is of a type that has that field

pub fn implementation_languages( self, implementation_languages: Vec<String>, ) -> Result<Self, Error>

Set the implementation_languages field for an SDO under construction, if the SDO is of a type that has that field

pub fn indicator_types(self, types: Vec<String>) -> Result<Self, Error>

Set the indicator_types field for an SDO under construction, if the SDO is of a type that has that field

pub fn infrastructure_types( self, infrastructure_types: Vec<String>, ) -> Result<Self, Error>

Set the infrastructure_types field for an SDO under construction, if the SDO is of a type that has that field

pub fn installed_software_refs( self, installed_software_refs: Vec<Identifier>, ) -> Result<Self, Error>

pub fn kill_chain_phases( self, phases: Vec<KillChainPhase>, ) -> Result<Self, Error>

Set the kill_chain_phases field for an SDO under construction, if the SDO is of a type that has that field

pub fn last_observed(self, last_observed: String) -> Result<Self, Error>

Set the last_observed field for an SDO under construction, if the SDO is of a type that has that field

pub fn last_seen(self, last_seen: String) -> Result<Self, Error>

Set the last_seen field for an SDO under construction, if the SDO is of a type that has that field

pub fn latitude(self, latitude: f64) -> Result<Self, Error>

pub fn longitude(self, longitude: f64) -> Result<Self, Error>

pub fn malware_types(self, malware_types: Vec<String>) -> Result<Self, Error>

Set the malware_types field for an SDO under construction, if the SDO is of a type that has that field

pub fn modules(self, modules: Vec<String>) -> Result<Self, Error>

pub fn name(self, name: String) -> Result<Self, Error>

Set the name field for an SDO under construction, if the SDO is of a type that has that field

pub fn number_observed(self, number_observed: u64) -> Result<Self, Error>

Set the number_observed field for an SDO under construction, if the SDO is of a type that has that field

pub fn objective(self, objective: String) -> Result<Self, Error>

pub fn object_refs(self, object_refs: Vec<Identifier>) -> Result<Self, Error>

Set the object_refs field for an SDO under construction, if the SDO is of a type that has that field

pub fn operating_system_ref( self, operating_system_ref: Identifier, ) -> Result<Self, Error>

Set the operating_system_ref field for an SDO under construction, if the SDO is of a type that has that field

pub fn operating_system_refs( self, operating_system_refs: Vec<Identifier>, ) -> Result<Self, Error>

Set the operating_system_ref field for an SDO under construction, if the SDO is of a type that has that field

pub fn opinion(self, opinion: OpinionType) -> Result<Self, Error>

Set the opinion field for an SDO under construction, if the SDO is of a type that has that field

pub fn pattern(self, pattern: String) -> Result<Self, Error>

Set the pattern field for an SDO under construction, if the SDO is of a type that has that field

pub fn pattern_type(self, pattern_type: String) -> Result<Self, Error>

Set the pattern_type field for an SDO under construction, if the SDO is of a type that has that field If the pattern_type is “stix”, this also sets the pattern_version field to the default STIX version of “2.1”

pub fn pattern_version(self, pattern_version: String) -> Result<Self, Error>

Set the pattern_version field for an SDO under construction, if the SDO is of a type that has that field

pub fn personal_motivations( self, personal_motivations: Vec<String>, ) -> Result<Self, Error>

Set the personal_motivations field for an SDO under construction, if the SDO is of a type that has that field

pub fn postal_code(self, postal_code: String) -> Result<Self, Error>

pub fn precision(self, precision: f64) -> Result<Self, Error>

pub fn primary_motivation( self, primary_motivation: String, ) -> Result<Self, Error>

Set the primary_motivation field for an SDO under construction, if the SDO is of a type that has that field

pub fn product(self, product: String) -> Result<Self, Error>

Set the product field for an MalwareAnalysis SDO under construction, if the SDO is of a type that has that field

pub fn published(self, published: Timestamp) -> Result<Self, Error>

Set the published field for an SDO under construction, if the SDO is of a type that has that field

pub fn region(self, region: String) -> Result<Self, Error>

pub fn report_types(self, report_types: Vec<String>) -> Result<Self, Error>

Set the report_type field for an SDO under construction, if the SDO is of a type that has that field

pub fn resource_level(self, resource_level: String) -> Result<Self, Error>

Set the resource_level field for an SDO under construction, if the SDO is of a type that has that field

pub fn result_name(self, result_name: String) -> Result<Self, Error>

pub fn roles(self, roles: Vec<String>) -> Result<Self, Error>

Set the roles field for an SDO under construction, if the SDO is of a type that has that field

pub fn sample_ref(self, sample_ref: Identifier) -> Result<Self, Error>

Set the sample_ref field for an SDO under construction, if the SDO is of a type that has that field

pub fn secondary_motivations( self, secondary_motivations: Vec<String>, ) -> Result<Self, Error>

Set the secondary_motivations field for an SDO under construction, if the SDO is of a type that has that field

pub fn sectors(self, identity_sector: Vec<String>) -> Result<Self, Error>

Set the sectors field for an SDO under construction, if the SDO is of a type that has that field

pub fn set_abstract(self, set_abstract: String) -> Result<Self, Error>

pub fn set_family(self, is_family: bool) -> Result<Self, Error>

pub fn set_result(self, result: String) -> Result<Self, Error>

pub fn set_version(self, version: String) -> Result<Self, Error>

Set the version field for an SDO under construction, if the SDO is of a type that has that field

pub fn sophistication(self, sophistication: String) -> Result<Self, Error>

Set the sophistication field for an SDO under construction, if the SDO is of a type that has that field

pub fn street_address(self, address: String) -> Result<Self, Error>

pub fn submitted(self, submitted: String) -> Result<Self, Error>

pub fn threat_actor_types( self, threat_actor_types: Vec<String>, ) -> Result<Self, Error>

Set the threat_actor_type field for an SDO under construction, if the SDO is of a type that has that field

pub fn tool_types(self, tool_types: Vec<String>) -> Result<Self, Error>

Set the tool_types field for an SDO under construction, if the SDO is of a type that has that field

pub fn tool_version(self, tool_version: String) -> Result<Self, Error>

Set the tool_version field for an SDO under construction, if the SDO is of a type that has that field

pub fn valid_from(self, valid_from: &str) -> Result<Self, Error>

Set the valid_from field for an SDO under construction, if the SDO is of a type that has that field

pub fn valid_until(self, valid_until: &str) -> Result<Self, Error>

Set the valid_until field for an SDO under construction, if the SDO is of a type that has that field

pub fn build(self) -> Result<DomainObject, Error>

Builds a new SDO, using the information found in the DomainObjectBuilder

This performs a final check that all required fields for a given SDO type are included before construction. This also runs the stick_check() validation method on the newly constructed SDO.