rust_stix2::domain_objects
Module vocab

Enumerations for standardizing and categorizing types for SDOs.

Structs

ArchitectureExecutionEnvsIter

An iterator over the variants of ArchitectureExecutionEnvs

AttackMotivationIter

An iterator over the variants of AttackMotivation

AttackResourceLevelIter

An iterator over the variants of AttackResourceLevel

ContextTypeIter

An iterator over the variants of ContextType

IdentitySectorsIter

An iterator over the variants of IdentitySectors

ImplementationLanguageIter

An iterator over the variants of ImplementationLanguage

IndicatorPatternTypeIter

An iterator over the variants of IndicatorPatternType

IndicatorTypeIter

An iterator over the variants of IndicatorType

InfrastructureTypeIter

An iterator over the variants of InfrastructureType

MalwareCapabilityIter

An iterator over the variants of MalwareCapability

MalwareResultIter

An iterator over the variants of MalwareResult

MalwareTypeIter

An iterator over the variants of MalwareType

OpinionTypeIter

An iterator over the variants of OpinionType

RegionIter

An iterator over the variants of Region

ReportTypeIter

An iterator over the variants of ReportType

ThreatActorRoleIter

An iterator over the variants of ThreatActorRole

ThreatActorSophisticationIter

An iterator over the variants of ThreatActorSophistication

ThreatActorTypeIter

An iterator over the variants of ThreatActorType

ToolTypeIter

An iterator over the variants of ToolType

Enums

ArchitectureExecutionEnvs

Represents the different processor architectures within the STIX framework.

AttackMotivation

Represents attack motivations used in Intrusion Set and Threat Actor SDOs within the STIX framework.

AttackResourceLevel

Represents the resource levels available to threat actors within the STIX framework.

ContextType

Represents different types of contexts within the STIX framework.

IdentitySectors

Represents various identity sectors within the STIX framework.

ImplementationLanguage

Represents different programming languages within the STIX framework.

IndicatorPatternType

This is a non-exhaustive, open vocabulary that covers common pattern languages and is intended to characterize the pattern language that the indicator pattern is expressed in within the STIX framework.

IndicatorType

Indicator type is an open vocabulary used to categorize Indicators.

InfrastructureType

Represents the type of infrastructure used in cyber attacks within the STIX framework.

MalwareCapability

Represents different malware capabilities within the STIX framework.

MalwareResult

Represents the common types of results from scanner or tool analysis process within the STIX framework.

MalwareType

Represents the different types and functions of malware within the STIX framework.

OpinionType

Represents the various types of opinions within the STIX framework.

Region

Represents the world regions based on the United Nations geoscheme within the STIX framework.

ReportType

Represents the various types of reports within the STIX framework.

ThreatActorRole

Represents the various roles of threat actors within the STIX framework.

ThreatActorSophistication

Represents the various levels of threat actor sophistication within the STIX framework.

ThreatActorType

Represents the various types of threat actors within the STIX framework.

ToolType

Represents the various ToolType of threat actors within the STIX framework.