rust_stix2::domain_objects::vocab
Enum MalwareCapability

pub enum MalwareCapability {

    AccessesRemoteMachines,
    AntiDebugging,
    AntiDisassembly,
    AntiEmulation,
    AntiMemoryForensics,
    AntiSandbox,
    AntiVm,
    CapturesInputPeripherals,
    CapturesOutputPeripherals,
    CapturesSystemStateData,
    CleansTracesOfInfection,
    CommitsFraud,
    CommunicatesWithC2,
    CompromisesDataAvailability,
    CompromisesDataIntegrity,
    CompromisesSystemAvailability,
    ControlsLocalMachine,
    DegradesSecuritySoftware,
    DegradesSystemUpdates,
    DeterminesC2Server,
    EmailsSpam,
    EscalatesPrivileges,
    EvadesAv,
    ExfiltratesData,
    FingerprintsHost,
    HidesArtifacts,
    HidesExecutingCode,
    InfectsFiles,
    InfectsRemoteMachines,
    InstallsOtherComponents,
    PersistsAfterSystemReboot,
    PreventsArtifactAccess,
    PreventsArtifactDeletion,
    ProbesNetworkEnvironment,
    SelfModifies,
    StealsAuthenticationCredentials,
    ViolatesSystemOperationalIntegrity,

}

Represents different malware capabilities within the STIX framework.

Variants

AccessesRemoteMachines

Indicates the ability to access remote machines.

AntiDebugging

Indicates the ability to prevent debugging or make it more difficult.

AntiDisassembly

Indicates the ability to prevent disassembly or make it more difficult.

AntiEmulation

Indicates the ability to prevent execution inside an emulator or make it more difficult.

AntiMemoryForensics

Indicates the ability to prevent or complicate memory forensics.

AntiSandbox

Indicates the ability to prevent sandbox-based analysis or make it more difficult.

AntiVm

Indicates the ability to prevent VM-based analysis or make it more difficult.

CapturesInputPeripherals

Indicates the ability to capture data from input peripherals, like keylogging.

CapturesOutputPeripherals

Indicates the ability to capture data from output peripherals, like screen scraping.

CapturesSystemStateData

Indicates the ability to capture information about a systemโs state.

CleansTracesOfInfection

Indicates the ability to clean traces of infection from a system.

CommitsFraud

Indicates the ability to commit fraud, such as click fraud.

CommunicatesWithC2

Indicates the ability to communicate with a command and control server.

CompromisesDataAvailability

Indicates the ability to compromise data availability.

CompromisesDataIntegrity

Indicates the ability to compromise data integrity.

CompromisesSystemAvailability

Indicates the ability to consume system resources, compromising availability.

ControlsLocalMachine

Indicates the ability to control the local machine.

DegradesSecuritySoftware

Indicates the ability to bypass or disable security software.

DegradesSystemUpdates

Indicates the ability to disable system updates and patches.

DeterminesC2Server

Indicates the ability to identify command and control servers.

EmailsSpam

Indicates the ability to send spam email messages.

EscalatesPrivileges

Indicates the ability to escalate privileges.

EvadesAv

Indicates the ability to evade antivirus detection.

ExfiltratesData

Indicates the ability to exfiltrate data.

FingerprintsHost

Indicates the ability to fingerprint the host system.

HidesArtifacts

Indicates the ability to hide artifacts like files and open ports.

HidesExecutingCode

Indicates the ability to hide executing code.

InfectsFiles

Indicates the ability to infect files on the system.

InfectsRemoteMachines

Indicates the ability to infect remote machines.

InstallsOtherComponents

Indicates the ability to install additional components.

PersistsAfterSystemReboot

Indicates the ability to persist after a system reboot.

PreventsArtifactAccess

Indicates the ability to prevent access to its artifacts.

PreventsArtifactDeletion

Indicates the ability to prevent deletion of its artifacts.

ProbesNetworkEnvironment

Indicates the ability to probe the network environment.

SelfModifies

Indicates the ability to modify itself.

StealsAuthenticationCredentials

Indicates the ability to steal authentication credentials.

ViolatesSystemOperationalIntegrity

Indicates the ability to compromise system operational integrity.