rust_stix2::domain_objects::vocab
Enum MalwareType

pub enum MalwareType {

    Adware,
    Backdoor,
    Bot,
    Bootkit,
    Ddos,
    Downloader,
    Dropper,
    ExploitKit,
    Keylogger,
    Ransomware,
    RemoteAccessTrojan,
    ResourceExploitation,
    RogueSecuritySoftware,
    Rootkit,
    ScreenCapture,
    Spyware,
    Trojan,
    Unknown,
    Virus,
    Webshell,
    Wiper,
    Worm,

}

Represents the different types and functions of malware within the STIX framework.

Variants

Adware

Adware: Software funded by advertising, may gather sensitive user information.

Backdoor

Backdoor: Allows an attacker to perform actions on a remote system.

Bot

Bot: Resides on an infected system, part of a botnet, monitors a backdoor for instructions.

Bootkit

Bootkit: Targets the Master Boot Record of the target computer.

Ddos

DDoS: Used to perform a distributed denial of service attack.

Downloader

Downloader: A small trojan file programmed to download and execute other files.

Dropper

Dropper: Deposits an enclosed payload (generally, other malware) onto the target computer.

ExploitKit

ExploitKit: A software toolkit to target common vulnerabilities.

Keylogger

Keylogger: Monitors keystrokes and records them or sends them back to a central point.

Ransomware

Ransomware: Encrypts files on a system, demanding ransom for access codes.

RemoteAccessTrojan

RemoteAccessTrojan: A trojan capable of controlling a machine through remote commands.

ResourceExploitation

ResourceExploitation: Steals a systemโs resources, such as a malicious bitcoin miner.

RogueSecuritySoftware

RogueSecuritySoftware: A fake security product demanding money to clean phony infections.

Rootkit

Rootkit: Hides its files or processes to conceal its presence and activities.

ScreenCapture

ScreenCapture: Captures images from the target systemโs screen for exfiltration.

Spyware

Spyware: Gathers information on a userโs system without their knowledge.

Trojan

Trojan: Malicious program used to hack into a computer by misleading users.

Unknown

Unknown: Not enough information to determine the type of malware.

Virus

Virus: Replicates by reproducing itself or infecting other programs.

Webshell

Webshell: Malicious script used to maintain persistent access on a compromised web application.

Wiper

Wiper: Deletes files or entire disks on a machine.

Worm

Worm: Self-replicating program that executes itself without user intervention.