rust_stix2
Module extensions

Contains the data structures and implementation for predefined Stix Cyber-observable object extensions.

These provide additional functionality and customization options for various SCOs

Structs

AlternateDataStreamType

Alternate Data Stream Type

ArchiveExtension

Archive File Extension

FileExtensionsIter

An iterator over the variants of FileExtensions

HttpRequestExtension

The HTTP request extension specifies a default extension for capturing network traffic properties specific to HTTP requests.

IcmpExtension

The ICMP extension specifies a default extension for capturing network traffic properties specific to ICMP.

NetworkTrafficExtensionsIter

An iterator over the variants of NetworkTrafficExtensions

NtfsExtension

NFTS Extension

PdfExtension

PDF File Extension

ProcessExtensionsIter

An iterator over the variants of ProcessExtensions

RasterExtension

Raster Image File Extension

SocketExtenion

The Network Socket extension specifies a default extension for capturing network traffic properties associated with network sockets.

TcpExtension

The TCP extension specifies a default extension for capturing network traffic properties specific to TCP.

UnixAccountExtension

The UNIX account extension specifies a default extension for capturing the additional information for an account on a UNIX system.

UserAccountExtensionsIter

An iterator over the variants of UserAccountExtensions

WindowsPEOptionalHeaderType

Windows™ PE Optional Header Type

WindowsPESectionType

Windows™ PE Section Type The Windows PE Section type specifies metadata about a PE file section

WindowsPebinaryExtension

Windows PE Binary File Extension

WindowsPebinaryTypeOvIter

An iterator over the variants of WindowsPebinaryTypeOv

WindowsProcessExtension

The Windows Process extension specifies a default extension for capturing properties specific to Windows processes.

WindowsServiceExtension

Windows Service Extension

Enums

ExifTag

An EXIF tag, which can be either a string or an integer representing EXIF metadata fields

FileExtensions

Possible extensions for File SCOs

NetworkTrafficExtensions

Possible extensions for NetworkTraffic SCOs

ProcessExtensions

Possible extensions for Process SCOs

SpecialExtensions

Known predefined extensions specific to particular SCOs

UserAccountExtensions

Possible extensions for UserAccount SCOs

WindowsPebinaryTypeOv

Possible types of a Windows PE binary, specifying a file extension for capturing properties specific to Windows portable executable (PE) files

Functions

check_extension

Validate a known predefined extension expressed as a raw extension dictionary