rust_stix2::extensions
Struct WindowsPEOptionalHeaderType

pub struct WindowsPEOptionalHeaderType {

    pub magic_hex: Option<String>,
    pub major_linker_version: Option<i64>,
    pub minor_linker_version: Option<i64>,
    pub size_of_code: Option<u64>,
    pub size_of_initialized_data: Option<u64>,
    pub size_of_uninitialized_data: Option<u64>,
    pub address_of_entry_point: Option<i64>,
    pub base_of_code: Option<i64>,
    pub base_of_data: Option<i64>,
    pub image_base: Option<i64>,
    pub section_alignment: Option<i64>,
    pub file_alignment: Option<i64>,
    pub major_os_version: Option<i64>,
    pub minor_os_version: Option<i64>,
    pub major_image_version: Option<i64>,
    pub minor_image_version: Option<i64>,
    pub major_subsystem_version: Option<i64>,
    pub minor_subsystem_version: Option<i64>,
    pub win32_version_value_hex: Option<String>,
    pub size_of_image: Option<u64>,
    pub size_of_headers: Option<u64>,
    pub checksum_hex: Option<String>,
    pub subsystem_hex: Option<String>,
    pub dll_characteristics_hex: Option<String>,
    pub size_of_stack_reserve: Option<u64>,
    pub size_of_stack_commit: Option<u64>,
    pub size_of_heap_reserve: Option<u64>,
    pub size_of_heap_commit: Option<u64>,
    pub loader_flags_hex: Option<String>,
    pub number_of_rva_and_sizes: Option<i64>,
    pub hashes: Option<Hashes>,

}

Windows™ PE Optional Header Type

The Windows PE Optional Header type represents the properties of the PE optional header. An object using the Windows PE Optional Header Type MUST contain at least one property from this type.

For more information see https://docs.oasis-open.org/cti/stix/v2.1/os/stix-v2.1-os.html#_29l09w731pzc

Fields
magic_hex: Option<String>

Specifies the hex value that indicates the type of the PE binary.

major_linker_version: Option<i64>

Specifies the linker major version number.

minor_linker_version: Option<i64>

Specifies the linker minor version number.

size_of_code: Option<u64>

Specifies the size of the code (text) section. If there are multiple such sections, this refers to the sum of the sizes of each section.

size_of_initialized_data: Option<u64>

Specifies the size of the initialized data section. If there are multiple such sections, this refers to the sum of the sizes of each section.

size_of_uninitialized_data: Option<u64>

Specifies the size of the uninitialized data section. If there are multiple such sections, this refers to the sum of the sizes of each section.

address_of_entry_point: Option<i64>

Specifies the address of the entry point relative to the image base when the executable is loaded into memory.

base_of_code: Option<i64>

Specifies the address that is relative to the image base of the beginning-of-code section when it is loaded into memory.

base_of_data: Option<i64>

Specifies the address that is relative to the image base of the beginning-of-data section when it is loaded into memory.

image_base: Option<i64>

Specifies the preferred address of the first byte of the image when loaded into memory.

section_alignment: Option<i64>

Specifies the alignment (in bytes) of PE sections when they are loaded into memory.

file_alignment: Option<i64>

Specifies the factor (in bytes) that is used to align the raw data of sections in the image file.

major_os_version: Option<i64>

Specifies the major version number of the required operating system.

minor_os_version: Option<i64>

Specifies the minor version number of the required operating system.

major_image_version: Option<i64>

Specifies the major version number of the image.

minor_image_version: Option<i64>

Specifies the minor version number of the image.

major_subsystem_version: Option<i64>

Specifies the major version number of the subsystem.

minor_subsystem_version: Option<i64>

Specifies the minor version number of the subsystem.

win32_version_value_hex: Option<String>

Specifies the reserved win32 version value.

size_of_image: Option<u64>

Specifies the size of the image in bytes, including all headers, as the image is loaded in memory.

size_of_headers: Option<u64>

Specifies the size of the image in bytes, including all headers, as the image is loaded in memory.

checksum_hex: Option<String>

Specifies the checksum of the PE binary.

subsystem_hex: Option<String>

Specifies the subsystem (e.g., GUI, device driver, etc.) that is required to run this image.

dll_characteristics_hex: Option<String>

Specifies the flags that characterize the PE binary.

size_of_stack_reserve: Option<u64>

Specifies the size of the stack to reserve, in bytes

size_of_stack_commit: Option<u64>

Specifies the size of the stack to commit, in bytes.

size_of_heap_reserve: Option<u64>

Specifies the size of the local heap space to reserve, in bytes.

size_of_heap_commit: Option<u64>

Specifies the size of the local heap space to commit, in bytes.

loader_flags_hex: Option<String>

Specifies the reserved loader flags.

number_of_rva_and_sizes: Option<i64>

Specifies the number of data-directory entries in the remainder of the optional header.

hashes: Option<Hashes>

Specifies any hashes that were computed for the optional header.