rust_stix2
Module json

Functions for deserializing STIX objects from JSON strings.

Functions

field_check

Compares fields of a deserialized struct with a JSON string, returning an error for unknown fields.

find_differences

Function to find differences in two vectors.

get_keys

Function to get the keys of any serializable struct passed in

get_object_type_from_json

Function to get the STIX object type from a generic STIX object JSON by matching the given type