rust_stix2
Module types

Core types for representing and manipulating STIX data structures.

Structs

ExtensionTypeIter

An iterator over the variants of ExtensionType

ExternalReference

An external reference that describes pointers to information outside STIX.

GranularMarking

The granular-marking type defines how the marking-definition object referenced by the marking_ref property or a language specified by the lang property applies to a set of content identified by the list of selectors in the selectors property.

Hashes

A STIX 2.1 compliant hash list, with a key/value pair identifying the hashing algorithm used and the hashed value.

Identifier

A STIX 2.1 compliant identifier that uniquely identifies a STIX Object.

KillChainPhase

Represents a phase in a kill-chain, i.e. one of the phases an attacker may undertake to achieve their objective.

LegalHashTypesIter

An iterator over the variants of LegalHashTypes

ReferenceUrl

A URL with one or more hashes for the contents of the URL.

ScoTypesIter

An iterator over the variants of ScoTypes

SdoTypesIter

An iterator over the variants of SdoTypes

SroTypesIter

An iterator over the variants of SroTypes

StixDictionary

A STIX 2.1 compliant dictionary that captures an set of key/value pairs.

StixMetaTypesIter

An iterator over the variants of StixMetaTypes

Timestamp

A custom Timestamp struct that holds a jiff::Timestamp.

Enums

DictionaryValue

Possible primitive dictionary values

ExtensionType

The Extensions Type enumeration used in the Extension SMO and the extensions common property.

LegalHashTypes

Refers to the valid hash algorithms like MD5, SHA-1, and SHA-256 used for file identification.

ScoTypes

A list of all STIX Cyber Observable Object (SCO) types found in the STIX 2.1 standard.

SdoTypes

A list of all STIX Domain Objects (SDO) types found in the STIX 2.1 standard.

SroTypes

A list of all STIX Relationship Object (SRO) types found in the STIX 2.1 standard.

StixMetaTypes

A list of all STIX Meta Object (SMO) types found in the STIX 2.1 standard.

Traits

Identified

A trait for objects that have an identifier, providing a method to access it.

Functions

get_extension_type

Gets the extension type for a general extension

get_field_by_name

Function to get the value of a struct’s field given the field name

get_object_type

Function to return the STIX Object type associated with the “type” value, or return “custom” if the type is not recognized

stix_case

Converts a raw string to kebab-case.